run-process: fix handling and docs for check-status
Enhance the handling of the check-status argument so that it is
possible to pass `#t` to get the default handling of the exit
status. This means that wrapper functions which want to sometimes
check the error status no longer have to re-implement
`check-process-success` which is not exported. Also support `#f` to
skip checking the exit status, this is simpler than passing to pass
`void` or worse still `true`. Update the documentation to reflect
these changes. Also update the information about the return value
of run-process as it does not return the exit status of the process.