Add tests for socket server
Closes #89.

Created a simple echo server for testing epoll or kqueue-based socket
servers, provide that the test script is run on linux or BSD,
respectively. Happy to add additional tests before merging if there
are any suggestions.

I wasn't quite sure of the best way to make use of cond-expand to take
care of the system-dependent aspects of these tests, so definitely
open to advice here, as well.

Tested on linux and OpenBSD.