Extend std/os/kqueue interface to accept flags, filter-flags
To acheive the edge-triggered behavior desired in the kqueue
socket-server, it is necessary to be able to set additional flags, as
well as filter flags to adjust the socket low watermark.

A kqueue-kevent-disable function is also added to mirror the behavior
of epoll-ctl-mod, used in the epoll socket server when one end of the
socket is closed. See link in #99.

It is also possible for kevent to return EINTR, so kevent is wrapped
in do-retry-nonblock.