Export low-level kqueue functions, use in kqueue socket server
By exposing the lower-level kevent functionality, it becomes possible
to reduce the number of syscalls when adding events to a kqueue. The
kevent-set! function provides a feature similar to the EV_SET macro,
allowing for the facile population of kevent structs.

The socket server is updated to use these in the add-socket and
close-socket handlers.