std/make: add back build-deps caching
Re-enable the build-deps feature, but make it more conservative
by storing the timestamps of the inputs and outputs.
Since the outputs are involved, store the cahce in the libdir-prefix.

Simplify std/build-script , yet at the same time, add the feature
that the setting flags can (and must, if non-default) be specified
in the call to defbuild-script.