avoid lock contention when there are many readers
Use two separate convars, one for readers and one for writers.
Readers wait on the rcv, while writes wait on the wcv.
When the read lock is relinquished, if there are writers waiting we
wake up exactly one with signal.
When the write lock is relinquished, if there are writers waiting we
wake up exactly one with signal; otherwise we wake up all readers.