Improve safety of BIO api
The bio extension methods _erroneously_ assume that the buffer is an
input-buffer; this is blatantly false with delimited input buffers,
and also alternative implementations of buffers if someone is so
inclined.

This fixes the problem by checking the exact type and only apply the
dangerous stuff if it is an exact instance; otherwise fallback to
generic methods. This applies for the inline methods and also
read-char/peek-char.

This in turn necessitated the introduction of a `put-back` method to
the BufferedReader interface, otherwise generic read-char and
peek-char implementations are impossible to implement correctly.
I reckon this will be generally useful for parsers, so it's not
abstraction leakage.

It also fixes the issue with get-buffer-output-* not flushing the
buffer before retrieving the output.