The Gerbil Universal Binary, aka BACH
This takes us halfway to deep integration; we now have a universal
gerbil binary, which links and initializes the expander/compiler and
it is capable of acting as any gerbil tool with a softlink.

We only depend on gsi in bootstreap (boot-gxi), which gets as far as
building bach. After that, the snake eats its tail.