compiler: optimize calls to not check for known procedures
This is an interesting optimization, that has measurable effects on
the build time (8-10% improvement in my build).

When the optimizer sees a call to a known (non gambit primitive)
procedure, with the correct arity, it optimizes the call to happen in
unsafe context. This means that at runtime the code won't check
whether it is actually a procedure (the compiler knows it is) and do a
direct call.