Docs: 404 links, `generic.md` headlines, misc (#1091)
- Reference - Prelude - Core Expander Syntax: the opening paragraph now
links to the reference of the `current-expander-context` instead of
404ing.
	- fixed comment spacing in the `core-expand` example.

- Reference - Standard Library - Generics: all the predefined types
starting with `<` failed to render on the site, this has been corrected
and they should now render correctly, including the sidebar content of
said types.


- Reference - Standard Library - Iterators: now correctly links to the
Guide explaining iterators.

- refer to "the guide" on in on the website itself as "Guide" with
uppercase G. This seemed to be a pattern in other parts of the docs.

- Reference - Standard Library - Syntax Parameters: link to the
`std/actor-v18/message.ss` instead of `actor`. `actor` does not exist,
only `actor-v18` and `actor-v13`. Is `v18` the correct one to link to?