New Separate Compilation Mode (#1200)
While running the r7rs benchmarks, it was observed in several programs
that separate compilation with libgerbil was regressing quite a bit.
After a weekend long investigation, the problem was narrowed down to the
incremental link with libgerbil.

Hence, this pr: we don't even build libgerbil anymore but we keep the
object files and link them as needed. This resolves the regression with
quite good results.
And since we don't really need --enable-shared by default now, it is
flipped to build static as default.

While at it, the compilation of binaries that share individual modules
got quite a bit faster, as it is parallelized and caches the object
files.