Typed gerbil improvements (#1206)
Fixes 3 issues/bugs in typed gerbil:
- references in class slot types would not expand; see the regression
test
- types for mutable bindings are now inferred, as the greatest commong
type of the original value and the mutation; this allows us to remove
some casts -- see the unit test and the changes in loader.
- (most important) specializer generation was broken by the introduction
of the implicit using for methods; this is now fixed and has a
regression test so that it doesn't happen again in the future.