Prevent invalid URL fragments (#1227)
Currently sidebar and header links are sometimes empty because of
VuePress's slugify rules. When the URL fragment is empty, the page has a
tendency to jump around while viewing it. For example "->" on
https://cons.io/reference/std/actor.html exhibits this behavior. This
makes viewing the docs occasionally painful, as the page may scroll to
the top while viewing it.

While fixing the issue, we might as well have nicer links also, so some
work has been done to make common characters and patterns have more
descriptive link names. For example:

* "->" becomes "to"
* "!" becomes "bang"
* "--..." becomes "dashes"