Add `exists?` method to `S3Bucket` interface (#1330)
This functionality might seem redundant, but there are some pretty
severe drawbacks to using existing options:

- Using `S3Bucket-get` works _alright_ as a proxy, especially if the
file doesn't exist (mapping the thrown error to `#f`). However, if the
file **does** exist, downloading the file to confirm its existence is a
waste, especially if the file is large
- Using `S3Bucket-list-objects` can work well, but only if the quantity
of objects is small and/or the list of objects rarely change (enabling
caching)

I tried to make the method name match existing conventions, but it might
create an unfortunate ambiguity: `S3Bucket-exists?` versus
`S3-bucket-exists?`. I am open to changing the name if a better
suggestion exists